+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- (react-native): Allow plugins to be set in the JS layer. [#1064](https://github.com/bugsnag/bugsnag-js/pull/1064)
 - (expo): Add support for Expo SDK v39 [#1052](https://github.com/bugsnag/bugsnag-js/pull/1052)
 
 ## 7.3.5 (2020-09-16)